system MPI
support the use of system MPI. ENABLE_MPICH variable
controls when we build MPICH internally. WITHOUT_MPI
variable controls if we build things that depend on
MPI. Adds a configuration file for mpi4py which can
be used to select between the internal MPICH build
and a number of others such as NERSC Edison.